requirements.txt
filelock==3.18.0
fsspec==2025.7.0
Jinja2==3.1.6
joblib==1.5.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.2
pandas==2.3.1
Pygments==2.19.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
rich==14.1.0
scikit-learn==1.7.1
scipy==1.16.1
six==1.17.0
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.7.1
typing_extensions==4.14.1
tzdata==2025.2
